refactor(EDCO-15): divided header in components
